feat: 방이 진행중인 경우 삭제할 수 없는 기능 추가

+++ src/main/java/chess/ChessService.java
@@ -26,9 +26,11 @@ import chess.utils.PieceFactory;
 @Service
 public class ChessService {
 
-    private static final String NO_ROOM_MESSAGE = "해당 ID와 일치하는 Room이 존재하지 않습니다.";
-    private static final String NO_SQUARE_MESSAGE = "해당 방, 위치에 존재하는 Square가 없습니다.";
-    private static final String NO_SQUARES_MESSAGE = "해당 ID에 체스게임이 초기화되지 않았습니다.";
+    private static final String NO_ROOM_MESSAGE = "[ERROR]: 해당 ID와 일치하는 Room이 존재하지 않습니다.";
+    private static final String NO_SQUARE_MESSAGE = "[ERROR]: 해당 방, 위치에 존재하는 Square가 없습니다.";
+    private static final String NO_SQUARES_MESSAGE = "[ERROR]: 해당 ID에 체스게임이 초기화되지 않았습니다.";
+    private static final String INVALID_PASSWORD = "[ERROR]: Password가 일치하지 않습니다.";
+    private static final String DELETE_NOT_ALLOWED_WHEN_RUNNING = "[ERROR]: 진행중인 방은 삭제할 수 없습니다.";
 
     private final RoomDao roomDao;
     private final SquareDao squareDao;
@@ -124,8 +126,13 @@ public class ChessService {
 
     public boolean delete(Long roomId, String password) {
         if (roomDao.findByIdAndPassword(roomId, password).isEmpty()) {
-            return false;
+            throw new IllegalArgumentException(INVALID_PASSWORD);
         }
+
+        if (!roomDao.findByIdAndPassword(roomId, password).get().getTurn().equals("empty")) {
+            throw new IllegalStateException(DELETE_NOT_ALLOWED_WHEN_RUNNING);
+        }
+
         squareDao.removeAll(roomId);
         roomDao.delete(roomId);
         return true;
